Use Vec for bootrom to avoid stack overflow
The use of arrays the size of the BOOTROM (over 2MB) causes overflow's
on Windows as the stack size is much smaller than on Linux.
By using `Vec` we avoid stack allocation and subsequently also a
potential overflow.

I have also removed the use of `Box` for the bootrom as it's not
necessary. I think the original idea was to prevent stack allocation
here but in reality `Rc` puts it on the stack when calling `Rc::new()`.